Fixed and added all GET Endpoints
